requirements.txt
asgiref==3.11.1
boto3==1.42.49
botocore==1.42.49
Django==6.0.2
django-storages==1.14.6
grep==0.3.2
gunicorn==25.1.0
jmespath==1.1.0
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
